Add Client Loom Module to build.gradle
Added the Client Loom Module to build.gradle to allow windchimes to be ran on a solo machine or a dedicated server.

This fixes https://github.com/NightEpiphany/WindchimesUnofficialContinuted/issues/1
