+++ build.gradle
@@ -18,6 +18,18 @@ repositories {
 	// for more information about repositories.
 }
 
+loom {
+    splitEnvironmentSourceSets()
+
+    mods {
+        "windchimes" {
+            sourceSet sourceSets.main
+            sourceSet sourceSets.client
+        }
+    }
+
+}
+
 dependencies {
 	// To change the versions see the gradle.properties file
 	minecraft "com.mojang:minecraft:${project.minecraft_version}"
