+++ src/client/java/com/moigferdsrte/windchimes/refine/bamboo/BambooWallChimeRenderer.java
@@ -1,4 +1,4 @@
-package com.moigferdsrte.windchimes.refine.iron;
+package com.moigferdsrte.windchimes.refine.bamboo;
 
 import com.moigferdsrte.windchimes.compact.ChimeBlockEntityRenderState;
 import com.moigferdsrte.windchimes.compact.ChimeModel;
@@ -6,34 +6,34 @@ import com.moigferdsrte.windchimes.refine.Refine;
 import com.moigferdsrte.windchimes.refine.RefineClient;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.minecraft.client.model.*;
-import net.minecraft.client.render.*;
+import net.minecraft.client.render.OverlayTexture;
+import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.block.entity.BlockEntityRenderer;
 import net.minecraft.client.render.block.entity.BlockEntityRendererFactory;
-import net.minecraft.client.render.block.entity.state.CampfireBlockEntityRenderState;
 import net.minecraft.client.render.command.ModelCommandRenderer;
 import net.minecraft.client.render.command.OrderedRenderCommandQueue;
 import net.minecraft.client.render.state.CameraRenderState;
-import net.minecraft.client.util.SpriteIdentifier;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.Objects;
-
 @Environment(EnvType.CLIENT)
-public class IronChimeRenderer implements BlockEntityRenderer<IronChimeBlockEntity, ChimeBlockEntityRenderState> {
+public class BambooWallChimeRenderer implements BlockEntityRenderer<BambooWallChimeBlockEntity, ChimeBlockEntityRenderState> {
 
     private final ChimeModel root;
 
-    public IronChimeRenderer(BlockEntityRendererFactory.Context ctx) {
+    public BambooWallChimeRenderer(BlockEntityRendererFactory.Context ctx) {
         this.root = new ChimeModel(ctx.getLayerModelPart(RefineClient.CHIME));
     }
 
     @Override
-    public void updateRenderState(IronChimeBlockEntity entity, ChimeBlockEntityRenderState state, float tickProgress, Vec3d cameraPos, @Nullable ModelCommandRenderer.CrumblingOverlayCommand crumblingOverlay) {
+    public ChimeBlockEntityRenderState createRenderState() {
+        return new ChimeBlockEntityRenderState();
+    }
+
+    @Override
+    public void updateRenderState(BambooWallChimeBlockEntity entity, ChimeBlockEntityRenderState state, float tickProgress, Vec3d cameraPos, @Nullable ModelCommandRenderer.CrumblingOverlayCommand crumblingOverlay) {
         BlockEntityRenderer.super.updateRenderState(entity, state, tickProgress, cameraPos, crumblingOverlay);
         state.ringTicks = entity.ringingTicks;
         state.strengthDivisor = entity.strengthDivisor;
@@ -44,16 +44,11 @@ public class IronChimeRenderer implements BlockEntityRenderer<IronChimeBlockEnti
     }
 
     @Override
-    public ChimeBlockEntityRenderState createRenderState() {
-        return new ChimeBlockEntityRenderState();
-    }
-
-    @Override
     public void render(ChimeBlockEntityRenderState state, MatrixStack matrices, OrderedRenderCommandQueue queue, CameraRenderState cameraState) {
         ChimeModel.ChimeModelState modelState = new ChimeModel.ChimeModelState(state.ringTicks, state.strengthDivisor, state.time, state.tickDelta, state.hanging);
-        RenderLayer renderLayer = RenderLayer.getEntityCutout(Identifier.of(Refine.MOD_ID, "textures/entity/chime/iron_chime.png"));
+        RenderLayer renderLayer = RenderLayer.getEntityCutout(Identifier.of(Refine.MOD_ID, "textures/entity/chime/bamboo_chime.png"));
         this.root.setAngles(modelState);
-        matrices.translate(0.5f, 1f, 0.5f);
+        matrices.translate(0.5f, 0f, 0.5f);
         queue.submitModel(
                 this.root,
                 modelState,
